PE-20 refactoring - move sql queries constant in a dedicated file - refactor parameter validation middlewares with TaskEither
